chore: ui/Label/stringSet to ts
