+++ ts_debt/source_files.yml
@@ -1,6 +1,5 @@
 (Wont do) ./_externals/lamejs/lame.all.js: lame.all.js
 ./ui/Label/index.jsx: index.jsx
-./ui/Label/stringSet.js: stringSet.js
 ./ui/Label/types.js: types.js
 ./ui/Label/utils.js: utils.js
 ./ui/LegacyChannelHeader/index.jsx: index.jsx
